JCF: first steps on performing Python bindings for the core ATLAS TDAQ objects
What's accomplished in this commit can be summarized by the (possibly
temporary) test script dal_testing.py It takes an XML file which I
constructed (dal_testing.data.xml) that defines an object of class
Partition which refers to a segment via the OnlineInfrastructure
relation. The segment refers to a pair of applications. Using a
function ("get_all_applications") modeled closely on the ATLAS TDAQ
dal repo's function of the same name (see its dal_algo.py file in the
tdaq-09-05-00 release) dal_testing.py will seek out the applications
and print their names to the screen.

Things to note:

-"ObjectLocator" fills the same role as ATLAS TDAQ's "Helper_class"; I
 feel the new name is more indicative of what the class does

-The schema file off of which the DAL is based, core.schema.xml, has
 to be located in two separate locations given the path assumptions
 currently in play. Clearly this approach will need to be improved.

-I'd have preferred to have called "dal_classes" just "dal" after the
 repo; unfortunately the config repo, like in the case of ATLAS TDAQ,
 already has its own Python module called "dal".