JCF: added in tests for Python bindings of Partition's get_log_directory and get_segment, as well as Component's get_parents
Note that the test *.data.xml file is dal_testing.data.xml, which I've
constructed from hand. There doesn't appear to be a file _existing_ in
the original ATLAS TDAQ repos which is intended to be fed into
algorithm_tests.py. There's some script which can generate it, but I
believe it relies on a lot of ATLAS-specific packages/assumptions.

One consequence is that while the tests of get_log_directory and
get_segment both pass, the test of get_parents is failing since I
haven't yet figured out how to write a component with parents into
dal_testing.data.xml. Regardless, the Python/C++ test infrastructure
is likely correct, and this is literally a case of the test not liking
the input since it can't meaningfully test off of it.